COMMIT 20.06.2019__23-12__RABOTAET!

+++ src/main/java/com/javamaster/dao/UserDAO.java
@@ -16,7 +16,7 @@ public interface UserDAO {
 
     void cleanTable() throws SQLException;
 
-    void deleteId(int id) throws SQLException;
+    void deleteId(long id) throws SQLException;
 
     void updateId(int id, String name, String login, String password) throws SQLException;
